+++ restli-client/src/main/java/com/linkedin/restli/client/FindRequest.java
@@ -43,7 +43,7 @@ public class FindRequest<T extends RecordTemplate>
 {
   private final CompoundKey _assocKey;
 
-  FindRequest(Map<String, String> headers,
+  public FindRequest(Map<String, String> headers,
               List<HttpCookie> cookies,
               Class<T> templateClass,
               ResourceSpec resourceSpec,