+++ restli-int-test/src/test/java/com/linkedin/restli/examples/TestParseqBasedFluentClientApiWithProjections.java
@@ -22,6 +22,10 @@ import com.linkedin.restli.client.ParSeqRestliClientBuilder;
 import com.linkedin.restli.client.ParSeqRestliClientConfigBuilder;
 import com.linkedin.restli.client.RestLiResponseException;
 import com.linkedin.restli.client.util.PatchGenerator;
+import com.linkedin.restli.common.BatchCollectionResponse;
+import com.linkedin.restli.common.BatchFinderCriteriaResult;
+import com.linkedin.restli.common.CollectionMetadata;
+import com.linkedin.restli.common.CollectionResponse;
 import com.linkedin.restli.common.CreateIdEntityStatus;
 import com.linkedin.restli.common.CreateIdStatus;
 import com.linkedin.restli.common.EntityResponse;
@@ -29,14 +33,24 @@ import com.linkedin.restli.common.IdEntityResponse;
 import com.linkedin.restli.common.PatchRequest;
 import com.linkedin.restli.common.UpdateEntityStatus;
 import com.linkedin.restli.examples.greetings.api.Greeting;
+import com.linkedin.restli.examples.greetings.api.GreetingCriteria;
 import com.linkedin.restli.examples.greetings.api.Tone;
 import com.linkedin.restli.examples.greetings.api.ValidationDemo;
 import com.linkedin.restli.examples.greetings.api.myItem;
 import com.linkedin.restli.examples.greetings.api.myRecord;
+import com.linkedin.restli.examples.greetings.client.AutoValidationWithProjection;
 import com.linkedin.restli.examples.greetings.client.AutoValidationWithProjectionFluentClient;
+import com.linkedin.restli.examples.greetings.client.Batchfinders;
+import com.linkedin.restli.examples.greetings.client.BatchfindersFluentClient;
+import com.linkedin.restli.examples.greetings.client.CreateGreeting;
 import com.linkedin.restli.examples.greetings.client.CreateGreetingFluentClient;
+import com.linkedin.restli.examples.greetings.client.Greetings;
 import com.linkedin.restli.examples.greetings.client.GreetingsFluentClient;
+import com.linkedin.restli.examples.greetings.client.ManualProjections;
 import com.linkedin.restli.examples.greetings.client.ManualProjectionsFluentClient;
+import com.linkedin.restli.examples.greetings.client.PagingMetadataProjections;
+import com.linkedin.restli.examples.greetings.client.PagingMetadataProjectionsFluentClient;
+import com.linkedin.restli.examples.greetings.client.PartialUpdateGreeting;
 import com.linkedin.restli.examples.greetings.client.PartialUpdateGreetingFluentClient;
 import com.linkedin.restli.server.validation.RestLiValidationFilter;
 
@@ -96,9 +110,9 @@ public class TestParseqBasedFluentClientApiWithProjections extends RestLiIntegra
   @Test
   public void testGetWithProjection() throws Exception
   {
-    ManualProjectionsFluentClient projections = new ManualProjectionsFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
+    ManualProjections projections = new ManualProjectionsFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
 
-    CompletionStage<Greeting> result = projections.get(1L, false,
+    CompletionStage<Greeting> result = projections.get(1L,
         optionalParams -> optionalParams.withMask(mask -> mask.withMessage()));
     CompletableFuture<Greeting> future = result.toCompletableFuture();
     Greeting greeting = future.get(5000, TimeUnit.MILLISECONDS);
@@ -110,7 +124,7 @@ public class TestParseqBasedFluentClientApiWithProjections extends RestLiIntegra
   @Test
   public void testBatchGetWithProjection() throws Exception
   {
-    GreetingsFluentClient greetings = new GreetingsFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
+    Greetings greetings = new GreetingsFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
 
     Set<Long> ids = Sets.newHashSet(Arrays.asList(1L, 2L, 3L));
     CompletionStage<Map<Long, EntityResponse<Greeting>>> result = greetings.batchGet(ids,
@@ -132,7 +146,7 @@ public class TestParseqBasedFluentClientApiWithProjections extends RestLiIntegra
   @Test
   public void testCreateAndGetWithProjection() throws Exception
   {
-    CreateGreetingFluentClient greetings = new CreateGreetingFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
+    CreateGreeting greetings = new CreateGreetingFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
 
     String msg = Double.toString(Math.random());
     CompletionStage<IdEntityResponse<Long, Greeting>> result = greetings.createAndGet(getGreeting(msg),
@@ -147,7 +161,7 @@ public class TestParseqBasedFluentClientApiWithProjections extends RestLiIntegra
   @Test
   public void testBatchCreateAndGetWithProjection() throws Exception
   {
-    CreateGreetingFluentClient greetings = new CreateGreetingFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
+    CreateGreeting greetings = new CreateGreetingFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
 
     String msg1 = Double.toString(Math.random());
     String msg2 = Double.toString(Math.random());
@@ -168,7 +182,7 @@ public class TestParseqBasedFluentClientApiWithProjections extends RestLiIntegra
   @Test
   public void testPartialUpdateAndGetWithProjection() throws Exception
   {
-    PartialUpdateGreetingFluentClient
+    PartialUpdateGreeting
         greetings = new PartialUpdateGreetingFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
     Greeting original = getGreeting();
     String message = "Edited message: fluent api test partialUpdateAndGet";
@@ -184,7 +198,7 @@ public class TestParseqBasedFluentClientApiWithProjections extends RestLiIntegra
   @Test
   public void testBatchPartialUpdateAndGetWithProjection() throws Exception
   {
-    PartialUpdateGreetingFluentClient greetings = new PartialUpdateGreetingFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
+    PartialUpdateGreeting greetings = new PartialUpdateGreetingFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
 
     Map<Long, PatchRequest<Greeting>> inputs = new HashMap<>();
     Greeting original = getGreeting();
@@ -205,16 +219,16 @@ public class TestParseqBasedFluentClientApiWithProjections extends RestLiIntegra
   @Test
   public void testGetAllWithFieldProjection() throws Exception
   {
-    GreetingsFluentClient greetings = new GreetingsFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
+    Greetings greetings = new GreetingsFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
 
     // Create some greetings with "GetAll" in message so they will be returned by getAll test method..
     CompletionStage<List<CreateIdStatus<Long>>> createResult = greetings.batchCreate(
         Arrays.asList(getGreeting("GetAll").setId(200L), getGreeting("GetAll").setId(201L)));
 
-    CompletionStage<List<Greeting>> result = createResult.thenCompose(ids -> greetings.getAll(
+    CompletionStage<CollectionResponse<Greeting>> result = createResult.thenCompose(ids -> greetings.getAll(
         optionalParams -> optionalParams.withMask(mask -> mask.withMessage())));
-    CompletableFuture<List<Greeting>> future = result.toCompletableFuture();
-    List<Greeting> greetingList = future.get(5000, TimeUnit.MILLISECONDS);
+    CompletableFuture<CollectionResponse<Greeting>> future = result.toCompletableFuture();
+    List<Greeting> greetingList = future.get(5000, TimeUnit.MILLISECONDS).getElements();
     Assert.assertTrue(greetingList.size() >= 2);
     for (Greeting greeting :greetingList)
     {
@@ -223,13 +237,208 @@ public class TestParseqBasedFluentClientApiWithProjections extends RestLiIntegra
     }
   }
 
+  @Test
+  public void testGetAllWithPagingProjection() throws Exception
+  {
+    PagingMetadataProjections
+        greetings = new PagingMetadataProjectionsFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
+
+    CompletionStage<CollectionResponse<Greeting>> result = greetings.getAll(
+        params -> params.withPagingMask(pageMask -> pageMask.withTotal().withStart())
+    );
+    CompletableFuture<CollectionResponse<Greeting>> future = result.toCompletableFuture();
+    CollectionResponse<Greeting> greetingResult = future.get(5000, TimeUnit.MILLISECONDS);
+    Assert.assertTrue(greetingResult.hasPaging());
+    Assert.assertFalse(greetingResult.getPaging().hasCount());
+    Assert.assertTrue(greetingResult.getPaging().hasStart());
+    Assert.assertEquals((int) greetingResult.getPaging().getStart(), 0);
+    Assert.assertFalse(greetingResult.getPaging().hasLinks());
+    Assert.assertTrue(greetingResult.getPaging().getTotal() > 0);
+
+    // Same request without projection.
+    result = greetings.getAll();
+    future = result.toCompletableFuture();
+    greetingResult = future.get(5000, TimeUnit.MILLISECONDS);
+    Assert.assertTrue(greetingResult.hasPaging());
+    Assert.assertTrue(greetingResult.getPaging().hasCount());
+    Assert.assertTrue(greetingResult.getPaging().hasStart());
+    Assert.assertTrue(greetingResult.getPaging().hasLinks());
+    // The resource applies manual projection and returns total only when it is explicitly projected. So no total field here.
+  }
+
+  @Test
+  public void testGetAllWithMetadataProjection() throws Exception
+  {
+    PagingMetadataProjections greetings = new PagingMetadataProjectionsFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
+
+    CompletionStage<CollectionResponse<Greeting>> result = greetings.getAll(
+        params -> params.withMetadataMask(metadataMask -> metadataMask.withMessage())
+    );
+    CompletableFuture<CollectionResponse<Greeting>> future = result.toCompletableFuture();
+    CollectionResponse<Greeting> greetingResult = future.get(5000, TimeUnit.MILLISECONDS);
+    Assert.assertNotNull(greetingResult.getMetadataRaw());
+    Greeting metadata = new Greeting(greetingResult.getMetadataRaw());
+    Assert.assertTrue(metadata.hasMessage());
+    Assert.assertFalse(metadata.hasTone());
+
+    // Same request without projection.
+    result = greetings.findByMetadataAutomaticPagingFullyAutomatic();
+    future = result.toCompletableFuture();
+    greetingResult = future.get(5000, TimeUnit.MILLISECONDS);
+    Assert.assertNotNull(greetingResult.getMetadataRaw());
+    metadata = new Greeting(greetingResult.getMetadataRaw());
+    Assert.assertTrue(metadata.hasMessage());
+    Assert.assertTrue(metadata.hasTone());
+  }
+
+  @Test
+  public void testFinderWithProjection() throws Exception
+  {
+    Greetings greetings = new GreetingsFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
+
+    CompletionStage<CollectionResponse<Greeting>> result = greetings.findBySearch(
+        params -> params.setTone(Tone.FRIENDLY)
+            .withMask(fieldsMask -> fieldsMask.withTone().withMessage())
+    );
+    CompletableFuture<CollectionResponse<Greeting>> future = result.toCompletableFuture();
+    List<Greeting> greetingList = future.get(5000, TimeUnit.MILLISECONDS).getElements();
+    Assert.assertTrue(greetingList.size() > 0);
+    for (Greeting greeting : greetingList)
+    {
+      Assert.assertEquals(greeting.getTone(), Tone.FRIENDLY);
+      Assert.assertFalse(greeting.hasId());
+      Assert.assertTrue(greeting.hasMessage());
+    }
+  }
+
+  @Test
+  public void testFinderWithPagingProjection() throws Exception
+  {
+    PagingMetadataProjections greetings = new PagingMetadataProjectionsFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
+
+    CompletionStage<CollectionResponse<Greeting>> result = greetings.findByMetadataAutomaticPagingFullyAutomatic(
+        params -> params.withPagingMask(pageMask -> pageMask.withTotal())
+    );
+    CompletableFuture<CollectionResponse<Greeting>> future = result.toCompletableFuture();
+    CollectionResponse<Greeting> greetingResult = future.get(5000, TimeUnit.MILLISECONDS);
+    Assert.assertTrue(greetingResult.hasPaging());
+    Assert.assertFalse(greetingResult.getPaging().hasCount());
+    Assert.assertFalse(greetingResult.getPaging().hasStart());
+    Assert.assertFalse(greetingResult.getPaging().hasLinks());
+    Assert.assertTrue(greetingResult.getPaging().hasTotal());
+    Assert.assertTrue(greetingResult.getPaging().getTotal() > 0);
+
+    // Same request without projection.
+    result = greetings.findByMetadataAutomaticPagingFullyAutomatic();
+    future = result.toCompletableFuture();
+    greetingResult = future.get(5000, TimeUnit.MILLISECONDS);
+    Assert.assertTrue(greetingResult.hasPaging());
+    Assert.assertTrue(greetingResult.getPaging().hasCount());
+    Assert.assertTrue(greetingResult.getPaging().hasStart());
+    Assert.assertTrue(greetingResult.getPaging().hasLinks());
+    Assert.assertTrue(greetingResult.getPaging().hasTotal());
+    Assert.assertTrue(greetingResult.getPaging().getTotal() > 0);
+  }
+
+  @Test
+  public void testFinderWithMetadataProjection() throws Exception
+  {
+    PagingMetadataProjections greetings = new PagingMetadataProjectionsFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
+
+    CompletionStage<CollectionResponse<Greeting>> result = greetings.findByMetadataAutomaticPagingFullyAutomatic(
+        params -> params.withMetadataMask(metadataMask -> metadataMask.withId())
+    );
+    CompletableFuture<CollectionResponse<Greeting>> future = result.toCompletableFuture();
+    CollectionResponse<Greeting> greetingResult = future.get(5000, TimeUnit.MILLISECONDS);
+    Assert.assertNotNull(greetingResult.getMetadataRaw());
+    Greeting metadata = new Greeting(greetingResult.getMetadataRaw());
+    Assert.assertFalse(metadata.hasMessage());
+    Assert.assertFalse(metadata.hasTone());
+    Assert.assertTrue(metadata.hasId());
+
+    // Same request without projection.
+    result = greetings.findByMetadataAutomaticPagingFullyAutomatic();
+    future = result.toCompletableFuture();
+    greetingResult = future.get(5000, TimeUnit.MILLISECONDS);
+    Assert.assertNotNull(greetingResult.getMetadataRaw());
+    metadata = new Greeting(greetingResult.getMetadataRaw());
+    Assert.assertTrue(metadata.hasMessage());
+    Assert.assertTrue(metadata.hasTone());
+    Assert.assertTrue(metadata.hasId());
+  }
+
+  @Test
+  public void testBatchFinderWithProjection() throws Exception
+  {
+    Batchfinders batchFinders = new BatchfindersFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
+    GreetingCriteria c1 = new GreetingCriteria().setId(1L).setTone(Tone.SINCERE);
+    GreetingCriteria c2 = new GreetingCriteria().setId(2L).setTone(Tone.FRIENDLY);
+    CompletionStage<BatchCollectionResponse<Greeting>> result = batchFinders.findBySearchGreetings(
+        Arrays.asList(c1, c2), "hello world",
+        params -> params.withMask(mask -> mask.withTone().withId()));
+    CompletableFuture<BatchCollectionResponse<Greeting>> future = result.toCompletableFuture();
+    List<BatchFinderCriteriaResult<Greeting>> batchResult = future.get(5000, TimeUnit.MILLISECONDS).getResults();
+
+    List<Greeting> greetings1 = batchResult.get(0).getElements();
+    Assert.assertTrue(greetings1.get(0).hasTone());
+    Assert.assertEquals(greetings1.get(0).getTone(), Tone.SINCERE);
+    Assert.assertTrue(greetings1.get(0).hasId());
+    Assert.assertFalse(greetings1.get(0).hasMessage());
+
+    List<Greeting> greetings2 = batchResult.get(1).getElements();
+    Assert.assertTrue(greetings2.get(0).hasId());
+    Assert.assertEquals(greetings2.get(0).getTone(), Tone.FRIENDLY);
+    Assert.assertTrue(greetings2.get(0).hasId());
+    Assert.assertFalse(greetings2.get(0).hasMessage());
+  }
+
+  @Test
+  public void testBatchFinderWithPagingProjection() throws Exception
+  {
+    Batchfinders batchFinders = new BatchfindersFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
+    GreetingCriteria c1 = new GreetingCriteria().setId(1L).setTone(Tone.SINCERE);
+    GreetingCriteria c2 = new GreetingCriteria().setId(2L).setTone(Tone.FRIENDLY);
+    CompletionStage<BatchCollectionResponse<Greeting>> result = batchFinders.findBySearchGreetings(
+        Arrays.asList(c1, c2), "hello world",
+        params -> params.withPagingMask(mask -> mask.withStart())
+            .withMask(fieldMask -> fieldMask.withId().withTone())
+    );
+    CompletableFuture<BatchCollectionResponse<Greeting>> future = result.toCompletableFuture();
+    List<BatchFinderCriteriaResult<Greeting>> batchResult = future.get(5000, TimeUnit.MILLISECONDS).getResults();
+
+    for (BatchFinderCriteriaResult<?> singleCriteria : batchResult)
+    {
+      CollectionMetadata paging = singleCriteria.getPaging();
+      Assert.assertTrue(paging.hasStart());
+      Assert.assertFalse(paging.hasCount());
+      Assert.assertFalse(paging.hasTotal());
+      Assert.assertFalse(paging.hasLinks());
+    }
+
+    // Same request without paging mask (setting field mask to avoid validation error on server)
+    result = batchFinders.findBySearchGreetings(Arrays.asList(c1, c2), "hello world",
+        params -> params.withMask(fieldMask -> fieldMask.withId().withTone())
+    );
+    future = result.toCompletableFuture();
+    batchResult = future.get(5000, TimeUnit.MILLISECONDS).getResults();
+
+    for (BatchFinderCriteriaResult<?> singleCriteria : batchResult)
+    {
+      CollectionMetadata paging = singleCriteria.getPaging();
+      Assert.assertTrue(paging.hasStart());
+      Assert.assertTrue(paging.hasCount());
+      Assert.assertTrue(paging.hasTotal());
+      Assert.assertTrue(paging.hasLinks());
+    }
+  }
+
   // Test Get request without projection
   @Test
   public void testGetFull() throws Exception
   {
-    ManualProjectionsFluentClient projections = new ManualProjectionsFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
+    ManualProjections projections = new ManualProjectionsFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
 
-    CompletionStage<Greeting> result = projections.get(1L, false);
+    CompletionStage<Greeting> result = projections.get(1L);
     CompletableFuture<Greeting> future = result.toCompletableFuture();
     Greeting greeting = future.get(5000, TimeUnit.MILLISECONDS);
     Assert.assertTrue(greeting.hasId());
@@ -241,10 +450,11 @@ public class TestParseqBasedFluentClientApiWithProjections extends RestLiIntegra
   @Test
   public void testGetWithProjectionDisabled() throws Exception
   {
-    ManualProjectionsFluentClient projections = new ManualProjectionsFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
+    ManualProjections projections = new ManualProjectionsFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
 
-    CompletionStage<Greeting> result = projections.get(1L, true,
-        optionalParams -> optionalParams.withMask(mask -> mask.withMessage()));
+    CompletionStage<Greeting> result = projections.get(1L,
+        optionalParams -> optionalParams.withMask(mask -> mask.withMessage())
+            .setIgnoreProjection(true));
     CompletableFuture<Greeting> future = result.toCompletableFuture();
     Greeting greeting = future.get(5000, TimeUnit.MILLISECONDS);
     // these fields would have been excluded by the framework if automatic projection was enabled
@@ -255,7 +465,7 @@ public class TestParseqBasedFluentClientApiWithProjections extends RestLiIntegra
   @Test
   public void testValidationWithNoProjection() throws Exception
   {
-    AutoValidationWithProjectionFluentClient
+    AutoValidationWithProjection
         validationDemos = new AutoValidationWithProjectionFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
     CompletionStage<ValidationDemo> result = validationDemos.get(1);
     try {
@@ -273,7 +483,7 @@ public class TestParseqBasedFluentClientApiWithProjections extends RestLiIntegra
   @Test
   public void testValidationWithOnlyValidFieldsProjected() throws Exception
   {
-    AutoValidationWithProjectionFluentClient validationDemos = new AutoValidationWithProjectionFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
+    AutoValidationWithProjection validationDemos = new AutoValidationWithProjectionFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
 
     CompletionStage<ValidationDemo> result = validationDemos.get(1,
         optionalParams -> optionalParams.withMask(mask -> mask.withStringB()